183 - adjusts  endpoint pages
